# soya-framework-admin/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "title" : "Workshop Server",
    "contact" : {
      "name" : "wenqun.soya@gmail.com"
    }
  },
  "basePath" : "/api",
...
  "schemes" : [ "http" ],
  "paths" : {
    "/ant/mkdir" : {
      "post" : {
        "tags" : [ "ant" ],
        "operationId" : "mkdir",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "dir",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/ant/build" : {
      "post" : {
        "tags" : [ "ant" ],
        "operationId" : "build",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "baseDir",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "buildFile",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "task",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/cmd/command/{group}/{command}" : {
      "get" : {
        "tags" : [ "Command Service" ],
        "operationId" : "command",
        "consumes" : [ "text/plain" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "group",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "command",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  }
}